# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook AutoPoster API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface FacebookPage {
  id: number;
  pageId: string;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  accessToken: string;
  createdAt: string;
}

export interface ConnectPageBody {
  /** Facebook user access token — we will fetch page details automatically */
  accessToken: string;
}

export type PostPostType = (typeof PostPostType)[keyof typeof PostPostType];

export const PostPostType = {
  text: "text",
  image: "image",
  video: "video",
} as const;

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus];

export const PostStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
  failed: "failed",
} as const;

export interface Post {
  id: number;
  pageId: number;
  /** @nullable */
  pageName?: string | null;
  /** @nullable */
  title?: string | null;
  postType: PostPostType;
  caption: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  status: PostStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  facebookPostId?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export type CreatePostBodyPostType =
  (typeof CreatePostBodyPostType)[keyof typeof CreatePostBodyPostType];

export const CreatePostBodyPostType = {
  text: "text",
  image: "image",
  video: "video",
} as const;

export interface CreatePostBody {
  pageId: number;
  /** @nullable */
  title?: string | null;
  postType?: CreatePostBodyPostType;
  caption: string;
  /** @nullable */
  imageUrl?: string | null;
  /**
   * AI prompt — if set and imageUrl is absent, the scheduler generates the image at publish time
   * @nullable
   */
  mediaPrompt?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}

export interface BulkCreatePostsBody {
  posts: CreatePostBody[];
}

export type UpdatePostBodyPostType =
  (typeof UpdatePostBodyPostType)[keyof typeof UpdatePostBodyPostType];

export const UpdatePostBodyPostType = {
  text: "text",
  image: "image",
  video: "video",
} as const;

export type UpdatePostBodyStatus =
  (typeof UpdatePostBodyStatus)[keyof typeof UpdatePostBodyStatus];

export const UpdatePostBodyStatus = {
  draft: "draft",
  scheduled: "scheduled",
} as const;

export interface UpdatePostBody {
  /** @nullable */
  title?: string | null;
  postType?: UpdatePostBodyPostType;
  caption?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  audioUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  status?: UpdatePostBodyStatus;
}

export interface GenerateImageBody {
  /** Text prompt to generate an image from */
  prompt: string;
  /**
   * Style hint e.g. photorealistic, cartoon, watercolor
   * @nullable
   */
  style?: string | null;
}

export interface GeneratedImage {
  imageUrl: string;
  prompt: string;
}

export interface DashboardStats {
  totalPages: number;
  totalPosts: number;
  scheduledPosts: number;
  publishedPosts: number;
  draftPosts: number;
  failedPosts: number;
}

export type ListPostsParams = {
  status?: ListPostsStatus;
  pageId?: number;
};

export type ListPostsStatus =
  (typeof ListPostsStatus)[keyof typeof ListPostsStatus];

export const ListPostsStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
  failed: "failed",
} as const;